% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consume.R
\name{consume}
\alias{consume}
\title{Consume an Arbitrary Tidier}
\usage{
consume(x, ...)
}
\arguments{
\item{x}{an object to tidy}

\item{...}{Arguments to pass to other tidy methods}
}
\description{
Tidy an object, and then return it as a data frame
}
