% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_method.R
\name{find_method}
\alias{find_method}
\title{Find Tidy Method}
\usage{
find_method(generic = c("tidy", "glance", "augment"), class)
}
\arguments{
\item{generic}{One of \code{c("tidy", "glance", "augment")}}

\item{class}{character vector of object classes.}
}
\description{
Find a tidy method for an object
}
